# swagger.yaml
openapi: 3.0.3
info:
  title: Girls Got Game API
  description: Basketball training application API for mobile and web clients
  version: 2.0.0
  contact:
    name: Girls Got Game API Support
    email: support@girlsgotgame.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.girlsgotgame.app/api
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /profiles/me:
    get:
      tags:
        - Profiles
      summary: Get current user profile
      description: Retrieve the profile information for the authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Profiles
      summary: Update current user profile
      description: Update the profile information for the authenticated user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfile'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profiles:
    post:
      tags:
        - Profiles
      summary: Create or update profile
      description: Create or update user profile during onboarding
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfile'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profiles/leaderboard:
    get:
      tags:
        - Profiles
      summary: Get leaderboard
      description: Retrieve the top players by points
      responses:
        '200':
          description: Leaderboard retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profiles/{id}:
    get:
      tags:
        - Profiles
      summary: Get profile by ID
      description: Retrieve public profile information for a specific user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profiles/players:
    get:
      tags:
        - Profiles
      summary: Get all player profiles
      description: Retrieve all players for parent dashboard
      responses:
        '200':
          description: Players retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicProfile'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Workout Endpoints
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        image:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        totalPoints:
          type: integer
          minimum: 0
        role:
          type: string
          enum: [player, parent]
        childId:
          type: string
          nullable: true
        isOnboarded:
          type: boolean
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
          default: false
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProfile:
      type: object
      properties:
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [player, parent]
          default: player
        childId:
          type: string
          format: uuid
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
        isOnboarded:
          type: boolean
          default: false
        totalPoints:
          type: integer
          minimum: 0
          default: 0
    UpdateProfile:
      type: object
      properties:
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [player, parent]
        childId:
          type: string
          format: uuid
        isOnboarded:
          type: boolean
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
    PublicProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        totalPoints:
          type: integer
        jerseyNumber:
          type: integer
          nullable: true
        role:
          type: string
          enum: [player, parent]
        createdAt:
          type: string
          format: date-time
    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        totalPoints:
          type: integer
        jerseyNumber:
          type: integer
          nullable: true
        avatarUrl:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid data provided
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...